requirements.txt
acicobra
pyyaml